:zap: Refactoring
