+++ src/main/java/info/itsthesky/disky/core/BotOptions.java
@@ -1,4 +1,4 @@
-package info.itsthesky.disky;
+package info.itsthesky.disky.core;
 
 import net.dv8tion.jda.api.JDABuilder;
 import net.dv8tion.jda.api.requests.GatewayIntent;
