+++ src/main/java/info/itsthesky/disky/core/ErrorHandler.java
@@ -1,4 +1,4 @@
-package info.itsthesky.disky;
+package info.itsthesky.disky.core;
 
 public interface ErrorHandler {
