+++ src/main/java/info/itsthesky/disky/core/Utils.java
@@ -1,4 +1,4 @@
-package info.itsthesky.disky;
+package info.itsthesky.disky.core;
 
 import org.bukkit.ChatColor;
